Reduced Array::m_width from size_t to uint8_t.

+++ src/realm/array.cpp
@@ -182,7 +182,7 @@ void Array::init_from_mem(MemRef mem) noexcept
     m_is_inner_bptree_node = get_is_inner_bptree_node_from_header(header);
     m_has_refs             = get_hasrefs_from_header(header);
     m_context_flag         = get_context_flag_from_header(header);
-    m_width                = get_width_from_header(header);
+    m_width                = uint8_t(get_width_from_header(header));
     m_size                 = get_size_from_header(header);
 
     // Capacity is how many items there are room for
